refactor(orchestration): extract CoordinatorRegistryService
Two methods, 111 body lines, moved verbatim. 2693 -> 2573 lines. This is the
first service to take a collaborator: registerExternalCoordinator reads three
in-memory maps that live on WorkerSessionManager, through the synchronous
accessors Task 9 added for exactly this. The facade passes the instance it
already built, not a new one.

The plan named the wrong collaborator -- its table says questionFlow, which
neither method touches -- and listed a logger port that does not exist. The
Pick is now the four ports the code calls.

buildCoordinatorRouteChatMetadata moves here as an unexported module function.
recordCoordinatorRouteContext is its only caller repo-wide; Task 10 left it in
the facade because none of *its* methods used it. Importing it back from
orchestration-service.ts was not an option: it is a value, so
verbatimModuleSyntax would not erase the import, and the facade imports this
service class as a value -- a real runtime cycle, the same trap Task 12 hit.

Both methods open kernel.mutate and call the three accessors from inside that
critical section. The accessors are synchronous and never open a mutate, which
is why the non-reentrancy guard does not fire. Nothing was added, removed, or
relocated.

Verified by mutation: passing a freshly constructed WorkerSessionManager to
this service turns three tests red -- the concurrency suite's
pendingLogicalTransportSessions assertion, the source-regex singleton backstop,
and, unexpectedly, the frozen oracle's "external coordinator registration
cannot race with a worker ensure side effect for the same handle". The 24
golden fixtures stay green, as designed: a split map changes no persisted state.
